Network nodes module (#21)
This PR adds network node module to pdanalytics.
The module has a crawler that transverses the dcrd network and records the information of connected nodes.
Node location is gotten from IP stack (the API access token must be added to the config file).
A central postgres package is added in this PR where all database interaction takes. Modules that depends on postgres database are updated to use this central package.